fix: fix solutions to NumberDivider and TimeConvertor

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        int divider = 5;
+        float divider = 5;
         System.out.println(toBeDivided/divider);
     }
 }
